+++ package.json
@@ -40,7 +40,7 @@
     "np": "^10.0.0",
     "strip-ansi": "6.0.1",
     "ts-morph": "16.0.0",
-    "typescript": "5.4.5",
+    "typescript": "5.2.2",
     "vitest": "^1.4.0"
   }
 }
